requirements.txt
livekit-agents~=1.0
livekit-plugins-openai~=1.0
livekit-plugins-cartesia~=1.0
livekit-plugins-deepgram~=1.0
livekit-plugins-silero~=1.0
livekit-plugins-turn-detector~=1.0
livekit-plugins-noise-cancellation>=0.2.0,<1.0.0
livekit-plugins-google>=0.1.0
python-dotenv~=1.0
torch>=2.0.0
flask~=2.3.0
aiohttp>=3.8.0
httpx>=0.24.0
gunicorn>=20.1.0
fastapi~=0.95
uvicorn~=0.22.0
livekit-api>=0.5.0
numpy>=1.21.0
websockets>=10.0